Change to DB

+++ ESBAdminDB/src/com/ibm/esbadmin/DeleteMQObject.java
@@ -63,9 +63,10 @@ public class DeleteMQObject extends HttpServlet {
 			Util newUtil = new Util();
 			try {
 
-				int qMgrID = Integer.parseInt(request.getParameter("qMgr")
-						.toString());
-
+				//int qMgrID = Integer.parseInt(request.getParameter("qMgr").toString());
+				
+				long qMgrID = Long.parseLong(request.getParameter("qMgr").toString());
+				
 				String usrQmgrQuery = "SELECT QSM_QMGR_PORT, QSM_QMGR_HOST, QSM_QMGR_CHL  FROM QMGR_MSTR "
 						+ "WHERE QSM_ID = " + qMgrID;
 				conn = newUtil.createConn();
